billing_agent/python_script/requirements.txt
pyyaml
google-cloud-storage
google-cloud-bigquery
pytz
google-cloud-monitoring